requirements.txt
flask
requests==2.28.2
appinsights
opencensus-ext-azure